:recycle: move tests in a dedicated module
